+++ packages/plugin-seo/src/fields/MetaImage.tsx
@@ -19,7 +19,7 @@ type MetaImageProps = UploadInputProps & {
 }
 
 export const MetaImage: React.FC<MetaImageProps> = (props) => {
-  const { name, fieldTypes, label, pluginConfig, relationTo } = props || {}
+  const { name, fieldTypes, label, pluginConfig, relationTo, required } = props || {}
 
   const field: FieldType<string> = useField(props as Options)
 
@@ -29,7 +29,7 @@ export const MetaImage: React.FC<MetaImageProps> = (props) => {
   const [fields] = useAllFormFields()
   const docInfo = useDocumentInfo()
 
-  const { setValue, showError, value } = field
+  const { setValue, showError, value, errorMessage } = field
 
   const regenerateImage = useCallback(async () => {
     const { generateImage } = pluginConfig
@@ -68,6 +68,18 @@ export const MetaImage: React.FC<MetaImageProps> = (props) => {
       >
         <div>
           {label && typeof label === 'string' && label}
+
+          {required && (
+            <span
+              style={{
+                marginLeft: '5px',
+                color: 'var(--theme-error-500)',
+              }}
+            >
+              *
+            </span>
+          )}
+
           {typeof pluginConfig.generateImage === 'function' && (
             <React.Fragment>
               &nbsp; &mdash; &nbsp;
@@ -110,6 +122,8 @@ export const MetaImage: React.FC<MetaImageProps> = (props) => {
           collection={collection}
           fieldTypes={fieldTypes}
           filterOptions={{}}
+          errorMessage={errorMessage}
+          required={required}
           label={undefined}
           name={name}
           onChange={(incomingImage) => {